Attempting to Add Presentation ID
Currently, there is no presentation id field that is properly populated when the user creates a presentation. This is important for being able to delete and view presentations. This commit is attempting to fix it, although testing has shown it has not been fixed yet. More work to be done.